feat: introduce debug shell
Currently Servers time out while chain'ing to http://169.254.169.254/latest/user-data.
It is also not possible to ping them, as that results ins a Request timeout for icmp_seq.

I therefore introduced a short timeout for entering into the manual iPXE Shell for debugging the environment - this might be removed later on.